refactor: temp base path k8s
