+++ agent-control/tests/common/mod.rs
@@ -1,5 +1,6 @@
 pub(super) mod agent_control;
 pub(crate) mod attributes;
+pub(super) mod base_paths;
 /// Includes a OpAMP mock server to test scenarios involving OpAMP.
 pub(super) mod effective_config;
 pub(super) mod global_logger;
